Convert street address to coordinates when needed; Renamed helper.js to stationService.js
